02.00.00-alpha2

+++ antarest/matrixstore/repository.py
@@ -116,13 +116,17 @@ class MatrixContentRepository:
             return None
 
         data = json.load(open(file))
-        return MatrixContent.from_dict(data)  # type: ignore
+        return MatrixContent.parse_obj(data)
+
+    def exists(self, id: str) -> bool:
+        file = self.bucket / id
+        return file.exists()
 
     def save(self, content: MatrixContent) -> str:
-        stringify = content.to_json()
+        stringify = content.json()
         h = MatrixContentRepository._compute_hash(stringify)
-        (self.bucket / h).write_text(stringify)
-
+        if not (self.bucket / h).exists():
+            (self.bucket / h).write_text(stringify)
         return h
 
     def delete(self, id: str) -> None:
